mlir2chainer: trivial fix

+++ mlir2chainer/mlir2chainer/tests/test_average_pooling_2d.py
@@ -25,7 +25,7 @@ def test_average_pooling_2d():
     result = mlir.MLIR('AveragePooling2D', 'mlir2chainer_test', 0.1, input_names, output_names, nodes, [function])
     result.dump('average_pooling_2d.mlir')
     
-    x = np.random.randn(2,3,4,5).astype(np.float32)
+    x = np.random.randn(2,3,4,5)
     ref = function.run(x)
     with chainer.using_config('train', False):
         m = mlir2chainer.ChainerNN('average_pooling_2d.mlir')
